Upgrade Django to version 4.1 and fix import error for url() function
This commit upgrades the project to Django version 4.1, ensuring compatibility with the latest features and improvements. Additionally, it addresses an import error related to the usage of the url() function, which has been moved from django.conf.urls to django.urls module since Django 3.1. The necessary changes have been made throughout the codebase to import the url() function from the correct module, resolving the import error encountered during the execution of 'collectstatic' command.